frontend/backend/requirements.txt
Flask==2.3.2
flask-jwt-extended==4.5.3
celery==5.4.0
redis==5.0.8
mysql-connector-python==8.4.0
requests==2.31.0
python-whois==0.8.0
pyOpenSSL==24.2.1
boto3==1.34.162
python-dotenv==1.0.1
flask-restx==1.3.0
gunicorn==22.0.0
slack-sdk==3.31.0